requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
decorator==5.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
proglog==0.1.12
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.57.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
